refactor(richmond-buildings): map Trazaloid openings like ArcFloor
Passages become centered lower-band doors (largest face extent wins per
side); door size comes from the AABB, not params. Apertures are not
mapped — the waist gap is the window. Floor/ceiling are `None|Solid`,
with Solid overridden by slab-cutting openings. Split geometry /
openings / slabs into submodules; update playground and connecting-shells.

## Agent Dialogue

- **Developer:** Give Trazaloid the ArcFloor openings treatment: passages → centered doors (largest wins; no door_* params); apertures unmapped (waist is the window); slabs `None|Solid` cut by non-door/aperture openings.
- **Agent:** Implemented submodule split (`geometry` / `openings` / `slabs`), passage resolution + slab cuts, dropped `door_*` / `SquareHole`, updated demos/playground/tests (all green).
