+++ maybraid/richmond/buildings/src/demos/connecting_shells.rs
@@ -92,9 +92,6 @@ impl ConnectingShells {
 				connect_id.clone(),
 				side_passage_opening(TrazaloidSide::West, footprint, 1.2, 2.1),
 			),
-			door_width_frac: 0.28,
-			door_thickness: 1.2,
-			door_height_frac: 0.7,
 			floor: TrazaloidSlab::None,
 			ceiling: TrazaloidSlab::Solid,
 			style: richmond_building_components::panels::PanelStyle::RoughStonework,
